Add feh and download some wallpapers
At the moment I just put feh in the windowmgr playbook but wallpaper
handlers might deserve their own playbooks in the future. To run it, I
patched the i3 config but I'd really like it to be separate from that so
that I don't have to worry about different config file syntax in case I
decide to change the wm.