Provision vmware vm with open-vm-tools-desktop
As it turns out, this is needed in order for copy/paste to work between
host and guest. Still, this doesn't fully work, I still have to manually
run the "vmtoolsd -n vmusr &" command in the terminal where I am trying
to copy/paste. I want to fix this properly somehow instead of adding
that command to an init script especially since it looks like systemd
already has a service running that; so for now live with the pain of
having to manually run vmtoolsd. There are some discussions here on
this issue: https://stackoverflow.com/questions/29647900/

Also increase the memory allocated to the vm.