Install nvim from source
The reason why I decided to do this is that one of the plugins pushed a
change which did not work with the nvim version installed from the
Ubuntu repositories.

So the choices were to pin said plugin to an earlier version or upgrade
nvim. I heard that this happens quite often in the nvim ecosystem so I
decided to go for the most recent stable nvim version in hopes that that
is the target that most plugins attempt to maintain compatibility with.